requirements.txt
numpy==2.0.2
pillow==10.4.0
streamlit==1.37.1
streamlit-navigation-bar==3.3.0
tensorflow==2.18.0